Version 24.3.22
[Engine]
* Use UL suffixed numbers instead of static_cast.
* Add cvar 'fmv_abuffer' to control maximum audio buffer size.
* Initialise maximum drift settings in Video class constructor and make it's setting non-volatile.
* Prefer 'size_t' for storing Video class IO buffering size and Video settings variables don't need to be atomic since the Video object construction only happens in the engine thread.
* Make sure video handling function suspends no matter what.
* ClockManagerType can be private and reorder functions in ClockInterval.
* Fix some AL function const correctness as alDelete* accepts const pointers.
* Don't need to allocate double memory when dealing with 16-bit video PCM conversion.
* Fixed miscalculations in FlagIsSet and FlagIsClear.
* Fix bad audio playback when there is no Theora stream by creating separate code paths for video only and audio only Oggs.
* Rewind a source when the source manager reuses it and use floating point values in its init function
* Random comment fixes.

[Assets]
* Update refdoc for new cvar.

+++ src/archive.hpp
@@ -472,7 +472,7 @@ static Archive *ArchiveInitNew(const string &strFile, const size_t stSize=0)
 /* -- Set extraction buffer size ------------------------------------------- */
 static CVarReturn ArchiveSetBufferSize(const size_t stSize)
   { return CVarSimpleSetIntNLG(cArchives->stExtractBufSize, stSize,
-      static_cast<size_t>(262144), static_cast<size_t>(16777216)); }
+      262144UL, 16777216UL); }
 /* -- Loads the archive from executable ------------------------------------ */
 static CVarReturn ArchiveInitExe(const bool bCheck)
 { // If we're checking the executable for archive?
